multi_lingual_story_telling/requirements.txt
streamlit>=1.30.0
openai>=1.0.0
gTTS>=2.4.0
PyPDF2>=3.0.0
pandas>=1.5.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
